diagnostics: Fix -fdump-internal-locations for 64-bit location_t
When adding support for 64-bit location_t in GCC 15, I missed a couple
changes needed for the internal debugging tool -fdump-internal-locations to
work properly. This would previously ICE on a location_t large enough to
overflow a signed 32-bit int.

gcc/ChangeLog:

	* diagnostic.cc (num_digits): Change argument type from `int' to
	`uint64_t'.
	(test_num_digits): Add test for 64-bit argument.
	* diagnostic.h (num_digits): Adjust prototype.
	* input.cc (write_digit): Accept argument in range 0-9 instead of
	an arbitrary int.
	(write_digit_row): Adjust to change in write_digit().

gcc/testsuite/ChangeLog:

	* gcc.dg/plugin/location-overflow-test-3.c: New test.
	* gcc.dg/plugin/plugin.exp: Add the new test.

+++ gcc/diagnostic.h
@@ -1254,7 +1254,7 @@ extern const char *diagnostic_get_color_for_kind (diagnostic_t kind);
 extern char *build_message_string (const char *, ...) ATTRIBUTE_PRINTF_1;
 
 /* Compute the number of digits in the decimal representation of an integer.  */
-extern int num_digits (int);
+extern int num_digits (uint64_t);
 
 inline bool
 warning_enabled_at (location_t loc, diagnostic_option_id option_id)
